# A few more pickings in the compiler perf opt