~3-4% improvement:

    compile_edgeql_script_simple: Mean +- std dev: 1.40 ms += 0.02 ms (5.45% faster)
    compile_edgeql_script_webapp_configure_system: Mean +- std dev: 18.8 ms += 0.3 ms (3.24% faster)
    compile_edgeql_script_webapp_get_movie: Mean +- std dev: 114 ms += 2 ms (2.44% faster)
    compile_edgeql_script_webapp_get_person: Mean +- std dev: 132 ms += 2 ms (2.48% faster)
    compile_edgeql_script_webapp_get_user: Mean +- std dev: 70.4 ms += 1.0 ms (3.47% faster)
    compile_edgeql_script_webapp_insert_movie: Mean +- std dev: 69.4 ms += 1.0 ms (3.94% faster)
    compile_edgeql_script_webapp_insert_person: Mean +- std dev: 19.6 ms += 0.3 ms (4.49% faster)
    compile_edgeql_script_webapp_insert_review: Mean +- std dev: 35.0 ms += 0.7 ms (3.57% faster)